refactor(core): extract common util methods
extract common util methods
